[Map] Voronoi seeds random generation
The class VoronoiSeedsGenerator has been implemented.

Its role is to randomly generate 2D points, which are going
to be Voronoi's diagram vertices.

In order to do so, we generate coodinates according to a
gaussian distribution, centered on the center of the map.

On top of that, so as to avoid "degenerated patterns",
possibly entailed by coinciding, or extremely closed, points.

Thus, we have subdivided the map thanks to a regular grid,
and imposed a maximum number of seeds by cell, plus a minimum
distance between two generated seeds.