State Machine Development
Beginning of the state machine development.
I think we need to change the conception of the state machine.
In each state we need information about the Boid and the list of other Boids
but with the design pattern state I think it's not possible to give these
argument because of a mutual inclusion of the different header.