[Map] Modification of VoronoiSeedsGenerator
The random seed generator has been modified so as to sort the
generated seeds according to their distances to the center of
the map.
This sort aims at making easier the "Whittaker Step".

On top of that, the .gitignore file has been modified in order
to ignore the compiled part of the library voro++.