[Bug Fixed] Modifying conception - Fixing Segfault
Due to the fact that the "voronoicell_neighbor" class implemented
in voro++ has no copy constructor appliable to constant cells, we had
to modify our conception choices.
As a matter of fact, formerly, our "Seed" class inherited from the
"voronoicell_neighbor" class. However, as previously said, the lack
of copy constructor for constant seeds was unsolvable.

In order to fix the situation, we chose to transform inheritance into
agregation (in other words, "Seed" know contains a "voronoicell_neighbor")
and we modified the voro++ library so as to provide a copy constructor for
constant seeds.
To be more precise, the agregation is realized thanks to a shared pointer
on a "voronoicell_neighbor".

Thus, all the classes integrating the "Seeds" had to be slighlty modified.