Implementation of the automaton
The automaton is now complete. It means that a boid go from one state to
another following some conditions. It needs now the implementation of
the force compute in each state and some condition (is a boid following
a prey ? is a boid next to a water source ? is a boid in love ?)