[Map] Computing and storing the centroids
So as to have the centers of the Voronoi cells, we compute the
centroid of each cell and store it in the "Seed" class.

Having the centroids' coordinates will be useful in many of the
next steps of the project. For instance, in order to size the
repulsion/attraction forces a biome applies on a boid (in other
word, if a rabbit's hole is in a biome, he will be attracted toward
this particular biome when he is out of it. If the seed is near the
border of the cell, the way the rabbit will be attracted toward the biome
may be anisotropic).