[Map] Modification of the comportment of the height tree
The tree do not use height blobs anymore. Indeed, the blending
of the height blobs, because they were placed on a grid, created
a pattern that made the grid appear in homogeneous area.

It has been replaced by a simple interpolation between the
height of the four vertices, keeping the multi-resolution
idea that is beneath the tree.