[Map - Fix] VoronoiSeedsGenerator list insertion
The function which aims at inserting a Seed, according to its distance to
the center of the map, in the list of generated seeds was incorrect.

In fact, the last test of the function, which aimed at determining if
either the end of the list was reached or a farer seed found, could
lead to a dereferencement of the "end pointer" of the iterator on the
list.

This flaw has been correctedd, and no more forbidden dereferencement of
the iterator on the list is done.