[Win Port] Modification of the CMakeLists.txt
So as to run CMake in order to build our project both for a Unix and a
Windows systems, the CMakeLists.txt file has been modified.

Mainly, all the dependance toward the "extlib" have been taken into
account thanks to conditionnal inclusion (CMake running either on a
Windows system or on a Unix one) of the needed libraries.

On top of that, the building mode is now taken into account too (by
default, the project is built in Debug mode).