[Win Port] Compiling voro++ on Windows system
The voro++ library does not include support for Windows in its standard
downlodable version.

As a consequence, I had to wrote a CMakeLists.txt file in order to build
it on a Windows system (with Visual Studio compiler).

On top of that, a file, generated while building the library, needed to be
modify in order to be compiled (G++ compiler seems to be less strict
concerning the namespaces).