[Map 3D] Multiplexing the 2D map with the HeightMap
First step of the 3D porting of the 2D map, we just take into account the
HeightMap for each vertex of the Voronoi Diagram.
On top of that, we added a Plane of equation z=0, so as to simulate a calm
sea (which becomes the reference when it comes to altitude).