[Map - Boids] Interface concerning Lake biomes
Two functions have been added to the MapGenerator class, so as to have
an easy access to the positions of the Lake biomes.

The first one, getLakes, allows the programmer to retrieve iterators on the
vector, stored in the MapGenerator class, containing the positions of the
centroids of the Lake biomes (the coordinates are stored within glm::vec2
since the "Boids Teams" prefers this data structure rather than Vertex2D).

The second one, given plane coordinates, "returns" the coordinates of the
centroid of the closest Lake biome (this function may need an update when
the boids will interact with the tridimensional map).