[Map - JSon] Creation of a class to parse parameters from a JSon file
In order to access and modify on-th-fly the simulation parameters (without having
to re-compile at each modification), the RapidJSon library has been included in
our branch of the project.

In a few words, all the previously "defines" parameters are going to be stored in a JSon
configuration file, and then parsed at initialisation of the project (in the main.ccp file).
That will have several positive consequences :
	- All the simulation parameters are going to be centralized in one single file ;
	- Modifying a parameter will not imly a re-compilation ;
	- All the simulation parameters' are going to be "global", since the "MapParameters"
		object will be accessible through the "MapGenerator" object.

In order to add a simulation parameter, the process will be the following :
	- Adding the parameter has an attribute of the MapParameter class ;
	- Adding the corresponding getter in the same class ;
	- Adding its parsing to the constructor of the same class ;
	- Giving to the class(es) which is(are) going to need this parameter a
		reference on the MapParameter object instanciated in the main.cpp
		file ;
	- Using the associated getter of the MapParameter object to get the
		value of the parameter.