[Map - JSon] Integration in Biome and BiomeRepartition
Since the changes involved by the integration of the JSon-parsed parameters are
deeply intricated, it is hard to properly segment the work done in indepedant commits...

That is why the commits to come are going to be quite "big", in terms of lines of code,
since I will only be able to commit when all the intricated changes of a set of changes
are done.

Here, the JSon-parsed parameters have been integrated in the two classes Biome and
BiomeRepartition.