[Map] Correction of the interpolation
The interpolation rules to compute the height of a position
within the height tree have been modified.
The previous formula created a asymmetry by deciding first
according to the first biome given as an argument.
The current formula is now symmetric regarding the two arguments.

However, the computation is here a bit slow as we compute for each
position 4 times a biome associated a location.
This will be soon corrected by a 'BiomeMap' which will enable
to have a fast (but approximative) a quick answer to theses requests.