[Map] Implementation of Lakes renderables
The lakes were not rendered so far (only the bottom texture was).
A class (LakeRenderable) dedicated to render lakes' water has been implemented but not yet tested
(because of the fact that we cannot execute our program through SSH/X).
On top of that, since lakes can be connexe, a computing phase was needed to determine
the outline of each sub-lake (in MapRenderable).

The tests are going to be done on a Windows system.