[Windows] Porting project to Windows systems
In order to complete the porting toward Windows systems, some changes had
to be made.

Indeed, the fact that we included the "extlib" repository into the project
(so as to ease its compilation, and, thus, its sharing/forking) compels us
to include both the Unix systems and the Windows systems version of the
former (the "extlib").

Consequently, the "extlib" architecture has been slightly modified:
	- The SFML, GLEW and TINYOBJLOADER libraries have been
	  "duplicated" and stored in two different repositories, namely
	  "Linux" and "Win32";
	- The Makefile aiming at building the "extlib" has been modified
	  so as to match the new architecture;
	- The CMakeList has also been modified according to the new
	  architecture;
	- The project file corresponding to the FREETYPE library had to be
	  updated so as to be used by Visual Studio 2015 (new version of
	  the C++ redistribuable library).

On top of that, since the Visual Studio 2015 compiler does not support
OpenMP 3.0, two source files had to be modified in order to reach
retro-compatibily (the changes are minor, and, thus, are not detailled
here... Just take a look at the "diff" if you are interested).