[Map] Height Map modified
Now all functions use solely a sampled height map
and not the height tree anymore.

For now, the height tree is kept to compute the
sampled height map, but the computation can now
be easily modified.

For a better interaction between the map and the
animals, one should increase the samping precision
(actually heightMapScaling in the JSON). This will
considerably increase the loading time (in
O(heightMapScaling^2)), but also obviously the
precision of the in heightMap.