[Map] Implementation of a parser for import/export
So as to allow to import and export maps, we had to determine the
necessary and sufficient map data to store/restore, namely:
	- The seeds used to build the underlaying Voronoi diagram (their
	  coordinates and biome types);
	- The sampled heightmap (the plane coordinates and the height of
	  each sampled point);
	- The JSon parameter file associated with the generated map to
	  import/export.

Consequently, specific file formats have been designed in order to
properly store the information to export/import (the former will be
detailedin a dedicated README.md file which will come with the next
commit).

And, in order to parse those specific file formats, a parser has been
implemented.
The main features of the latter consists in:
	- Skipping efficiently white-space characters (spaces,
	  tabulations, line feeds and carriage returns);
	- Parsing (and skipping, according to a boolean parameter)
	  sequences of characters according to their beginning characters
	  (quite useful for comments);
	- Parsing an integer, a float, a string (delimited by white-space
	  characters) and a seed;
	- Keeping track of the current parsed line.