[Map] Import/export - mapData updated
In order to store the exported maps, the intern structure of the directory
"mapData" has been updated.

Since the parameter file will have to be changed for each import (since
the latter must be in suit with the data to import), a copy of the
original parameter file is available in the directory "Original".

On top of that, a map, specific to a test situation involving lake
connexity, has been designed and is available in the "LakeConnexityTest"
directory.

Finally, the import/export feature, from its global architecture to its
details, is explained in the "README.md" at the root level of "mapData".