[Map] Fixing lake connexity issue
As it was precised in the global "README.md" file, some connexe "Lake"
biomes were not detected as connexe, resulting in weird looking
"two-plane" lakes.

As a matter of fact, the problem occured because we only fused the newly
added "Lake" biome to the first connexe lake found in the list of
connexe lakes.

The issue has been corrected by looking within all the connexe lakes if a
new connexity exists, and, if so, by fusing them all together.